+++ ReleaseNotes.md
@@ -7,6 +7,11 @@
 * SBBBubbleView with new arrow styling (featuring a circular border around the arrow).
 * SBBSegmentedPicker has now border around selected segment in dark mode.
 * New semantic SBBColor .border.
+* New semantic SBBColor .tabViewBackground replaces .controlBackground.
+* New official SBBColors .charcoal and .shadow.
+
+### Bugfixes
+* Updated background colors of SBBBubbleView and SBBSegmentedPicker
 
 ## V0.0.13 - 31 Mar 2020